switch to the AVX2 implementation of BLAKE2b
From https://github.com/sneves/blake2-avx2. As with BLAKE3, all upstream
files are vendored verbatim. I've added a separate blake2b.o target in
the Makefile that includes the necessary -mavx2 flag.

This implementation only exposes an all-at-once API, so it's not
well-suited for production as-is. But for this benchmark specifically,
all-at-once is fine, and expanding the API to include an incremental
hasher wouldn't meaningfully affect performance.